requirements.txt
datasets==2.16.1
language_tool_python==2.7.3
nemo_toolkit==1.21.0
noisereduce==3.0.2
numpy==1.26.4
pandas==1.5.3
PyAudio==0.2.14
pydub==0.25.1
scikit_learn==1.4.0
scipy==1.13.0
speechbrain==1.0.0
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
transformers==4.36.2
TTS==0.22.0
TTS==0.22.0
webrtcvad==2.0.10
~ransformers==4.38.0.dev0
